GitHub #300: Allow standard Maven build with any JDK.
Test cases which require source level > 1.5 get separate source folders.
These source folders are only considered if the corresponding
"bytecode.version" property is specified for integration tests.

The standard JaCoCo build targets Java 1.5 and will not execute any of
these tests (regardless of the JDK version used for building).
