+++ org.jacoco.core.test/src/org/jacoco/core/test/validation/Source.java
@@ -12,6 +12,7 @@
 package org.jacoco.core.test.validation;
 
 import java.io.BufferedReader;
+import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
 import java.io.Reader;
@@ -31,16 +32,18 @@ import java.util.regex.Pattern;
 public class Source {
 
 	/**
-	 * Reads the source for the given type from the <code>./src/</code> folder
-	 * relative to the working directory.
+	 * Reads the source for the given type from the given source folder relative
+	 * to the working directory.
 	 * 
+	 * @param srcFolder
+	 *            source folder
 	 * @param type
 	 *            type to load the source file for
-	 * @throws IOException
-	 * @throws
 	 */
-	public static Source getSourceFor(final Class<?> type) throws IOException {
-		String file = "src/" + type.getName().replace('.', '/') + ".java";
+	public static Source getSourceFor(final String srcFolder,
+			final Class<?> type) throws IOException {
+		File folder = new File(srcFolder);
+		File file = new File(folder, type.getName().replace('.', '/') + ".java");
 		return new Source(new FileReader(file));
 	}
